requirements.txt
torch
Pillow
numpy
opencv-python